Extract word lists from alphabet-view into separate components
