--- src/assets/variables.css
@@ -1,1 +1,0 @@
-/* variables */
